Upgrade DEB package standard and dependencies to »Trixie«
- compat level now managed with the new debhelper-compat scheme
- Trixie corresponds to version 13
- go through the Debian stardard updates for 3.9.6 -> 4.7.2
  https://www.debian.org/doc/debian-policy/upgrading-checklist.html#version-4-7-2

The package relied on CDBS for building, but that framework has
effectively be discontinued in Debian, while a majority of packages
has switched to the dh-Sequencer.

Debian clearly dislikes "strange" build systems, and so we are
on our own with `SCons`, yet `dh` should provide the necessary hooks,
and our build does already provide all the features to support packaging.