Replace CDBS by the dh-sequencer
Even while SCons is considered a "strange build system" by Debian,
achieving the same level of functionality with more "common" build systems
would require a massive amount of work or would not be possible at all.

CSBS had built-in support for SCons,
but with the dh-sequencer it is rather straight-forward to invoke
our buildsystem explicitly.