Build: provide and install a lumiera.desktop file
The XDG Desktop spec is accepted universally, and thus
the old Debian-specific 'menu' system is deprecated and
no longer engaged automatically.

See: https://lists.debian.org/debian-devel-announce/2015/09/msg00000.html

Thus we'll now provide a Desktop file and install that already from out
SCons build system, together with a suitable variation of the Lumiera icon.

TODO: not sure if everything was done the correct way
 * do we need to ''register'' the new file in some way (preinst script?)
 * the menu entry shows up, but not the icon
 * but if we put an absolute path for the lumiera.svg into the desktop file, it shows up

Remark: in later experiments with package building,
the menu entry and the icon showed up in the menu.
Not sure if this requires a reboot or some similar
trigger (like restart of the destkop)