Maintain copyright information in machine readable format
As recommended in the Debian Policy §12.5.1
https://www.debian.org/doc/debian-policy/ch-docs.html#machine-readable-copyright-information

we maintain a compilation of relevant copyright and licensing information
in the new formalised format. We apply the policy to cover the code base
as a whole with a copyright assigned to the "core developers" and
use a common file stanza with the compiled authorship information
and time ranges of copyright. Detailed information is added into
the copyright headers of individual code files.

As before, the License for Lumiera code is GPL-2 or later.

Content on the Website is additionally dual-licensed GPL-2 / CC-By-SA 4,
but this is not relevant for the DEB package.