Disable known fragile tests for DEB package
Some of the scheduler tests are notoriously fragile,
since they verify observable timings.

I have not yet found a better way to verify their objectives
and I could not determine timings that work reliably even directly
after a full-load package build.

Furthermore, a check is disabled, which verifies read-only state of a file;
this test fails when running as root, and this is commonplace when building
the DEB package in a container.

Maybe we need to relegate those tests into a special testsuite anyway?